rev-parse: fix meaning of rev~ vs rev~0.
I think it would make more sense for rev~ to have the same guarantees that
rev^ has, namely to always return a commit. I would also suggest that not
giving a number would have the same effect of defaulting to 1, not 0.

Right now it's a bit illogical, but at least it's an _undocumented_
illogical behaviour.

This patch makes '^' and '~' act the same for the default count (i.e. both
default to 1), and also have the same behaviour for a count of zero.

Before (no discernible pattern):

	[torvalds@woody git]$ git rev-parse v1.5.1 v1.5.1^0 v1.5.1~0 v1.5.1^ v1.5.1~
	45354a57ee7e3e42c7137db6c94fa968c6babe8d
	90d5888abb2e67ae93d2b4893d9649a431c41037
	45354a57ee7e3e42c7137db6c94fa968c6babe8d
	c8f989335f51d44b9ff9cb0dd4bd14f3d9c3579c
	45354a57ee7e3e42c7137db6c94fa968c6babe8d

After (fairly logical):

	[torvalds@woody git]$ git rev-parse v1.5.1 v1.5.1^0 v1.5.1~0 v1.5.1^ v1.5.1~
	45354a57ee7e3e42c7137db6c94fa968c6babe8d
	90d5888abb2e67ae93d2b4893d9649a431c41037
	90d5888abb2e67ae93d2b4893d9649a431c41037
	c8f989335f51d44b9ff9cb0dd4bd14f3d9c3579c
	c8f989335f51d44b9ff9cb0dd4bd14f3d9c3579c

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>